move code to existing rebuild.py
